refactor: remove _args prefix from modules in node_core
