+++ crates/node-core/src/args/pruning.rs
